SPARK-1779.  Allow error on parameters that violate logical bounds.
Add optional minValue/maxValue arguments to getInt, getLong and getDouble.  If either or both of these are provided, the given bound is checked and an error results if the current parameter setting is outside the bound(s).

Another optional boolean argument, defaultBadValue, will cause the configured value to take on the given default value if the bounds are violated, with a warning log message.  This argument defaults to false, in which case an exception will be thrown instead.   The defaultBadValue parameter is supported by getInt, getLong, getDouble and getBoolean.